Refactor into client and executors and achieve full run
