+++ cs_publish/__init__.py
@@ -1,7 +1,3 @@
-from cs_publish.publish import Publisher
+__version__ = "0.0.0.0406.0"
 
-__version__ = "0.0.0.0418.0"
-
-__ALL__ = [
-    "Publisher",
-]
+__ALL__ = []
