+++ cs_publish/client/publish.py
@@ -1,98 +1,18 @@
 import argparse
 import copy
-import json
 import os
 import sys
-import uuid
 import yaml
 from pathlib import Path
 
-import requests
-from git import Repo, InvalidGitRepositoryError
-from kubernetes import client as kclient, config as kconfig
+from ..utils import run, clean
 
-
-from cs_publish.utils import clean, run, parse_owner_title, read_github_file
-from cs_publish.secrets import Secrets
+from .core import Core
 
 TAG = os.environ.get("TAG", "")
 PROJECT = os.environ.get("PROJECT", "cs-workers-dev")
 CURR_PATH = Path(os.path.abspath(os.path.dirname(__file__)))
-BASE_PATH = CURR_PATH / ".."
-
-
-class Core:
-    cr = "gcr.io"
-
-    def __init__(self, project, tag=None, base_branch="origin/master", quiet=False):
-        self.tag = tag
-        self.project = project
-        self.base_branch = base_branch
-        self.quiet = quiet
-
-    def get_config(self, models):
-        config = {}
-        for owner_title in models:
-            owner, title = parse_owner_title(owner_title)
-            if (owner, title) in config:
-                continue
-            else:
-                config_file = (
-                    BASE_PATH / Path("config") / Path(owner) / Path(f"{title}.yaml")
-                )
-                if config_file.exists():
-                    with open(config_file, "r") as f:
-                        c = yaml.safe_load(f.read())
-                else:
-                    config_file = self.get_config_from_remote([(owner, title)])
-                config[(c["owner"], c["title"])] = c
-        if not self.quiet and config:
-            print("# Updating:")
-            print("\n#".join(f"  {o}/{t}" for o, t in config.keys()))
-        elif not self.quiet:
-            print("# No changes detected.")
-        return config
-
-    def get_config_from_diff(self):
-        try:
-            r = Repo()
-            files_with_diff = r.index.diff(r.commit(self.base_branch), paths="config")
-        except InvalidGitRepositoryError:
-            files_with_diff = []
-        config = {}
-        for config_file in files_with_diff:
-            with open(config_file.a_path, "r") as f:
-                c = yaml.safe_load(f.read())
-            config[(c["owner"], c["title"])] = c
-        return config
-
-    def get_config_from_remote(self, models):
-        config = {}
-        for owner_title in models:
-            owner, title = parse_owner_title(owner_title)
-            resp = requests.get(
-                "https://api.github.com/repos/compute-tooling/compute-studio-publish/contents/config/{owner}/{title}.yaml"
-            )
-            content = read_github_file(
-                "compute-tooling", "compute-studio", "master", f"{owner}/{title}.yaml"
-            )
-            config[(owner, title)] = yaml.safe_load(content)
-        return config
-
-    def _resources(self, app, action=None):
-        if action == "io":
-            resources = {
-                "requests": {"cpu": 0.7, "memory": "0.25G"},
-                "limits": {"cpu": 1, "memory": "0.7G"},
-            }
-        else:
-            resources = {"requests": {"memory": "1G", "cpu": 1}}
-            resources = dict(resources, **copy.deepcopy(app["resources"]))
-        return resources
-
-    def _list_secrets(self, app):
-        secret = Secrets(app["owner"], app["title"], self.project)
-        return secret.list_secrets()
+BASE_PATH = CURR_PATH / ".." / ".."
 
 
 class Publisher(Core):
@@ -108,7 +28,7 @@ class Publisher(Core):
 
     """
 
-    kubernetes_target = CURR_PATH / Path("..") / Path("kubernetes")
+    kubernetes_target = BASE_PATH / Path("kubernetes")
 
     def __init__(
         self,
@@ -119,7 +39,7 @@ class Publisher(Core):
         quiet=False,
         kubernetes_target=None,
     ):
-        super().__init__(self, tag, project, models, base_branch, quiet)
+        super().__init__(project, tag, base_branch, quiet)
 
         self.models = models if models and models[0] else None
         self.kubernetes_target = kubernetes_target or self.kubernetes_target
@@ -134,22 +54,15 @@ class Publisher(Core):
             self.config.update(self.get_config(self.models))
 
         with open(
-            CURR_PATH
-            / Path("..")
-            / Path("templates")
-            / Path("sc-deployment.template.yaml"),
-            "r",
+            BASE_PATH / Path("templates") / Path("sc-deployment.template.yaml"), "r"
         ) as f:
             self.app_template = yaml.safe_load(f.read())
 
-        with open(
-            CURR_PATH / Path("..") / Path("templates") / Path("job.template.yaml"), "r"
-        ) as f:
+        with open(BASE_PATH / Path("templates") / Path("job.template.yaml"), "r") as f:
             self.job_template = yaml.safe_load(f.read())
 
         with open(
-            CURR_PATH / Path("..") / Path("templates") / Path("secret.template.yaml"),
-            "r",
+            BASE_PATH / Path("templates") / Path("secret.template.yaml"), "r"
         ) as f:
             self.secret_template = yaml.safe_load(f.read())
 
@@ -295,6 +208,15 @@ class Publisher(Core):
         container_config["env"].append(
             {"name": "APP_NAME", "value": f"{safeowner}_{safetitle}_tasks"}
         )
+        container_config["env"].append(
+            {
+                "name": "REDIS",
+                "valueFrom": {
+                    "secretKeyRef": {"name": "worker-secret", "key": "REDIS"}
+                },
+            }
+        )
+
         self._set_secrets(app, container_config)
 
         if self.kubernetes_target == "-":
@@ -329,85 +251,6 @@ class Publisher(Core):
             )
 
 
-class Job(Core):
-    def __init__(self, project):
-        super().__init__(project, quiet=True)
-        self.config = {}
-        kconfig.load_kube_config()
-        self.api_client = kclient.BatchV1Api()
-        self.job = None
-
-    def env(self, owner, title, config):
-        safeowner = clean(owner)
-        safetitle = clean(title)
-        envs = [
-            kclient.V1EnvVar("OWNER", config["owner"]),
-            kclient.V1EnvVar("TITLE", config["title"]),
-            kclient.V1EnvVar("SIM_TIME_LIMIT", str(config["sim_time_limit"])),
-        ]
-
-        for secret in self._list_secrets(config):
-            envs.append(
-                kclient.V1EnvVarSource(
-                    secret_key_ref=(
-                        kclient.V1SecretKeySelector(
-                            key=secret, name=f"{safeowner}-{safetitle}-secret"
-                        )
-                    )
-                )
-            )
-        return envs
-
-    def configure(self, owner, title, tag, job_id=None):
-        if job_id is None:
-            job_id = str(uuid.uuid4())
-
-        if (owner, title) not in self.config:
-            self.config.update(self.get_config([(owner, title)]), remote=True)
-
-        config = self.config[(owner, title)]
-
-        safeowner = clean(owner)
-        safetitle = clean(title)
-        name = f"{safeowner}-{safetitle}"
-        job_name = f"{name}-{job_id}"
-        container = kclient.V1Container(
-            name=job_name,
-            image=f"{self.cr}/{self.project}/{safeowner}_{safetitle}_tasks:{tag}",
-            command=["cs-jobs", "--job-id", job_id],
-            env=self.env(owner, title, config),
-        )
-        # Create and configurate a spec section
-        template = kclient.V1PodTemplateSpec(
-            metadata=kclient.V1ObjectMeta(labels={"app": f"{name}", "job-id": job_id}),
-            spec=kclient.V1PodSpec(restart_policy="Never", containers=[container]),
-        )
-        # Create the specification of deployment
-        spec = kclient.V1JobSpec(template=template, backoff_limit=4)
-        # Instantiate the job object
-        job = kclient.V1Job(
-            api_version="batch/v1",
-            kind="Job",
-            metadata=kclient.V1ObjectMeta(name=job_name),
-            spec=spec,
-        )
-
-        if not self.quiet:
-            print(yaml.dump(job.to_dict()))
-
-        self.job = job
-
-    def create(self):
-        return self.api_client.create_namespaced_job(body=self.job, namespace="default")
-
-    def delete(self):
-        return self.api_client.delete_namespaced_job(
-            name=self.job.metadata.name,
-            namespace="default",
-            body=kclient.V1DeleteOptions(),
-        )
-
-
 def main():
     parser = argparse.ArgumentParser(description="Deploy C/S compute cluster.")
     parser.add_argument("--tag", required=False, default=TAG)
@@ -439,7 +282,3 @@ def main():
         publisher.push()
     if args.app_config:
         publisher.write_app_config()
-
-
-if __name__ == "__main__":
-    main()
