+++ cs_publish/utils.py
@@ -1,6 +1,9 @@
+import base64
 import re
 import subprocess
 import time
+
+import requests
 
 
 def clean(word):
@@ -28,13 +31,6 @@ def read_github_file(org, repo, branch, filename):
     """
     Read data from github api. Ht to @andersonfrailey for decoding the response
     """
-    # TODO: incorporate master branch
-    if "http" in filename:
-        raise URLFormatError(
-            "A URL was entered for a 'github_file' type attribute. "
-            "For more information, check out the catalog configuration "
-            "docs: Tools/Catalog-Builder/README.md"
-        )
     url = f"https://api.github.com/repos/{org}/{repo}/contents/{filename}?ref={branch}"
     response = requests.get(url)
     print(f"GET: {url} {response.status_code}")
