+++ scripts/celery_io.sh
@@ -2,4 +2,4 @@
 SAFEOWNER=$(python -c "import re, os; print(re.sub('[^0-9a-zA-Z]+', '', \"$1\").lower())")
 SAFETITLE=$(python -c "import re, os; print(re.sub('[^0-9a-zA-Z]+', '', \"$2\").lower())")
 cs-publish --help
-celery -A cs_publish.tasks worker --loglevel=info --concurrency=1 -Q ${SAFEOWNER}_${SAFETITLE}_inputs_queue -n ${SAFEOWNER}_${SAFETITLE}_inputs@%h
+celery -A cs_publish.executors.tasks worker --loglevel=info --concurrency=1 -Q ${SAFEOWNER}_${SAFETITLE}_inputs_queue -n ${SAFEOWNER}_${SAFETITLE}_inputs@%h
